move string utilties to localstack.utils.strings
